+++ app/controllers/patients/comparisons_controller.rb
@@ -20,6 +20,14 @@ module Patients
       @patient.execute_data_sync(current_user)
       pipelines.reload
 
+      @sources = []
+      pipelines.first.computations.select(&:revision).each do |compared_comp|
+        compare_to_comp = pipelines.second.computations.select(&:revision).
+                          detect { |c| c.pipeline_step == compared_comp.pipeline_step }
+
+        @sources << [compared_comp, compare_to_comp] if compare_to_comp
+      end
+
       @data = { compared: [], not_comparable: [] }
       pipelines.first.data_files.each do |compared|
         compare_to = pipelines.second.data_files.detect { |df| df.similar?(compared) }
@@ -36,7 +44,8 @@ module Patients
     private
 
     def pipelines
-      @pipelines ||= patient.pipelines.where(iid: params[:pipeline_ids]).includes(:data_files)
+      @pipelines ||= patient.pipelines.where(iid: params[:pipeline_ids]).
+                     includes(:data_files, :computations)
     end
 
     def patient