+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Data sets API reference added to the Help section as a separate entry (@dharezlak)
 - Use action cable (web sockets) to refresh computation (@mkasztelnik)
 - Computation stores information about selected tag/branch and revision (@mkasztelnik)
+- Pipelines comparison shows sources diff link to GitLab (@Nuanda)
 
 ### Changed
 - Upgrade to ruby 2.4.1 (@mkasztelnik)